// old/interfaces/customers.interfaces.ts
export interface Customer {
  id: string;
  name: string;
  email: string;
  cpf: string;
  password: string;
  phoneNumber: string;
  userId: string;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  cpf?: string;
  password?: string;
  phoneNumber?: string;
}

export interface CustomerRepository {
  findAll(): Promise<Customer[]>;
  findById(id: string): Promise<Customer | null>;
  update(id: string, data: CustomerUpdate): Promise<Customer | null>;
  delete(id: string): Promise<void>;
}